HS-95-04: embeddable page cores — one core, two hosts
pages/cores/ is the pattern's home: a core is host-agnostic (no router
hooks, no page chrome, scope by prop — tests/unit/test_page_cores_guard.py
locks it mechanically), owns its verbs, and hands them to a hero
render-slot so each host chooses the chrome. ActivityPage and CommandsPage
become thin wrappers around ActivityCore/CommandsCore; the desk hosts the
same cores through SurfaceWindows (one SURFACES row per surface: shell
key, window id, lazy core, unmountOnMinimize) with a left-flank default
home that respects the z ladder (chrome sits UNDER the window band — the
walk caught a centered default burying the Tools shelf). The tool shelf's
mechanism converts to the shell dispatcher: registered surfaces open
in-world, unregistered ones fall back to their legacy route until
HS-95-08 retires the fallbacks. The add-a-surface authoring note ships in
web/README.md.

Proof (evidence-story-04.md + assets/cores-1440.png): cores guard green,
web suite 257/257 (core-in-window + wrapper chrome tests; the Phase 93
arrival test updated to the dispatcher mechanically), and the cores walk
on the production bundle — shelf opens Activity and Commands as chrome-
free desk windows while /activity and /commands keep their hero.

PMO-Story: HS-95-04
PMO-Contract-Digest: sha256:42a3e7b08a2bbd91da1d30a0e3934f0e9a51de4c91caa1ebe76c1af3f414fd1b
